adicionando arquivos PID

+++ src/pid/pid.c
@@ -18,8 +18,8 @@ void pid_configura_constantes(double Kp_, double Ki_, double Kd_){
     Kd = Kd_;
 }
 
-void pid_atualiza_referencia(float referencia_){
-    referencia = (double) referencia_;
+void pid_atualiza_referencia(double referencia_){
+    referencia = referencia_;
 }
 
 double pid_controle(double saida_medida){
